refactor(stock): extract stockOrderService + idempotent PO write-offs
- evaluate/receiveIntoStock/findOrCreateSubstituteStock/resolveOrCreateStockItem
  move from routes/stockOrders.js to services/stockOrderService.js
  (behavior-preserving; ADR-0002 absorption math untouched)
- PO evaluate write-offs: awaited + ADR-0003 marker idempotency
  (stockLossRepo.noteMarkerExists) — no more duplicate loss rows on
  Eval Error retry, no more silent write-off loss
- converts the it.todo retry-idempotency gap into real integration tests
