+++ CLAUDE.md
@@ -117,7 +117,7 @@ These bug patterns have been found and fixed. Follow these rules to avoid reintr
 6. **PO lines need identity** — every PO line must have either a Stock Item link or a Flower Name. Validate on creation, not during evaluation.
 7. **Order Termination flows behind a shared seam — never inline.** Both Cancellation (Status → Cancelled, optional Stem return) and Deletion (record removed, Owner-only) live behind `useOrderTerminationFlow` (`packages/shared/hooks/`) + `OrderTerminationConfirm` (`packages/shared/components/`). Hosts pass `apiClient`, `showToast`, `t`, and `onSuccess`; the hook owns the two-button confirm state, the three endpoints (POST `/orders/:id/cancel-with-return`, PATCH `/orders/:id`, DELETE `/orders/:id`), and toast composition. Any new site that needs to cancel or delete an Order MUST consume the hook + component, never reimplement the handlers inline. Translation keys (`cancelConfirm`, `cancelAndReturn`, `cancelNoReturn`, `orderCancelled`, `orderDeleted`, `stockReturned`, `confirmDelete`) are exercised by the hook directly; per-app `translations.js` must define them. History: pre-2026-05-02 the florist app silently patched Status (stock decremented forever); 2026-05-09 the three lockstep sites (`apps/florist/src/components/OrderCard.jsx`, `apps/florist/src/pages/OrderDetailPage.jsx`, `apps/dashboard/src/components/OrderDetailPanel.jsx`) were migrated behind the seam — drift is now structurally impossible.
 8. **Stock / Committed is NOT a subtraction** — `Current Quantity` is decremented immediately on order creation (`orderService.js` → `atomicStockAdjust`). Every pending order's demand is therefore ALREADY reflected in `Current Quantity`. `GET /stock/committed` returns the SAME demand as an informational breakdown (for traceability and the tap-to-expand orders list) — it is NOT a second number to subtract. The correct formula is `effective = qty`. Always use `getEffectiveStock(qty)` from `packages/shared/utils/stockMath.js`, never inline `qty - committed` anywhere. A negative `qty` means genuine shortfall — buy more stems. If you need to understand drift between `qty` and physical reality, look at the `/stock/:id/usage` trace; do not invent formula tweaks. Any stock-math change must be audited on BOTH `apps/florist/src/components/StockItem.jsx` AND `apps/dashboard/src/components/StockTab.jsx`. History: pre-2026-04-22 used `qty - committed` (double-counted), 2026-04-22 first-attempted fix used `qty < 0 ? qty : qty - committed` (broke cumulative shortfall case), 2026-04-22 final fix collapsed to `effective = qty`.
-9. **A new Batch MUST carry its Variety attrs** — any code path that creates a Stock Item from a PO receipt (`receiveIntoStock` in `backend/src/routes/stockOrders.js`, and the sibling absorption site in `backend/src/routes/stockPurchases.js`) MUST write `Type / Colour / Size / Cultivar` onto the new dated Batch, sourced from the PO line and falling back to the orig Stock Item. When the orig Demand Entry has `type_name = NULL`, back-fill it in the SAME transaction. Why: under `STOCK_Y_MODEL=true`, `stockRepo.listGroupedByVariety` filters `type_name IS NOT NULL`, so an attr-less Batch is invisible in the grouped Stock view, FEFO can't compute its Variety key, and the legacy picker routes onto the undated orig instead. This was the root cause of #323 (stems "disappeared" after a PO evaluation). Absorption math (`batchQty = qty + existingQty` when negative, ADR-0002) is deliberate and unchanged — do NOT touch it while fixing attrs. Fixed in #327; regression-locked by `backend/src/__tests__/stockOrders.receiveIntoStock.integration.test.js`.
+9. **A new Batch MUST carry its Variety attrs** — any code path that creates a Stock Item from a PO receipt (`receiveIntoStock` in `backend/src/services/stockOrderService.js` — moved from `routes/stockOrders.js` in the W2 extraction, 2026-07-07, behavior unchanged, and the sibling absorption site in `backend/src/routes/stockPurchases.js`) MUST write `Type / Colour / Size / Cultivar` onto the new dated Batch, sourced from the PO line and falling back to the orig Stock Item. When the orig Demand Entry has `type_name = NULL`, back-fill it in the SAME transaction. Why: under `STOCK_Y_MODEL=true`, `stockRepo.listGroupedByVariety` filters `type_name IS NOT NULL`, so an attr-less Batch is invisible in the grouped Stock view, FEFO can't compute its Variety key, and the legacy picker routes onto the undated orig instead. This was the root cause of #323 (stems "disappeared" after a PO evaluation). Absorption math (`batchQty = qty + existingQty` when negative, ADR-0002) is deliberate and unchanged — do NOT touch it while fixing attrs. Fixed in #327; regression-locked by `backend/src/__tests__/stockOrders.receiveIntoStock.integration.test.js`.
 
 ## Key Files
 - `BACKLOG.md` — feature tracking, open items, known issues
