+++ backend/CLAUDE.md
@@ -24,7 +24,7 @@ scripts/           → Backfill, shadow-health, start-test-backend, etc.
 | customers.js | GET/POST/PATCH /customers, GET /customers/insights | CRM + RFM segmentation |
 | stock.js | GET/POST/PATCH /stock, GET /stock/velocity, GET /stock/committed, GET /stock/pending-po, POST /stock, GET /stock/:id/usage, GET /stock/varieties/:key/usage, PATCH /stock/:id, POST /stock/:id/adjust, POST /stock/:id/write-off, GET /stock/reconciliation, GET /stock/needs-backfill, GET /stock/distinct/:column, PATCH /stock/:id/variety-attrs, PATCH /stock/variety-attrs/bulk (Owner-only) | Inventory + velocity + variety backfill + per-Variety trace (T5) |
 | deliveries.js | GET/PATCH /deliveries | Driver assignments + status cascade to orders |
-| stockOrders.js | Full CRUD /stock-orders | PO lifecycle: create, send, shop, review, evaluate. **Postgres** via `stockOrderRepo` (Phase 7). New PO markers embed human-readable PO number (ADR-0003). |
+| stockOrders.js | Full CRUD /stock-orders | PO lifecycle: create, send, shop, review, evaluate. **Postgres** via `stockOrderRepo` (Phase 7). New PO markers embed human-readable PO number (ADR-0003). Routes are thin controllers — `resolveOrCreateStockItem` and the `/evaluate` handler delegate to `services/stockOrderService.js` (W2 extraction). |
 | stockPurchases.js | POST /stock-purchases | Record supplier deliveries with batch tracking. Routes STOCK reads/writes through stockRepo; purchase records through stockPurchasesRepo. |
 | stockLoss.js | GET/POST /stock-loss | Waste logging by reason |
 | premadeBouquets.js | CRUD /premade-bouquets | Premade bouquet templates (composed of stock items). Service uses `premadeBouquetRepo` (Postgres, Phase 7). |
@@ -50,6 +50,7 @@ scripts/           → Backfill, shadow-health, start-test-backend, etc.
 | File | Purpose |
 |------|---------|
 | orderService.js | Order state machine, auto-match stock, create with rollback, cancel with stock return, edit bouquet lines. All paths via orderRepo (Postgres). |
+| stockOrderService.js | PO evaluate flow — `evaluatePurchaseOrder(poId, lines)` (extracted from `routes/stockOrders.js`, W2), `receiveIntoStock` (dated-Batch receive w/ ADR-0002 absorption math + Variety-attr propagation, #327), `findOrCreateSubstituteStock` (substitute stock card creation, C13), `resolveOrCreateStockItem` (PO-line auto-link/create). Returns a domain result (`{outcome: 'conflict'\|'partial'\|'complete', ...}`) the route maps to HTTP — route stays a thin controller. Idempotency (ADR-0003 marker convention) covers BOTH receives (`stockPurchasesRepo.noteMarkerExists`) AND write-offs (`stockLossRepo.noteMarkerExists`, added alongside this extraction) — a write-off failure now marks the line `error` → PO → Eval Error → retry re-attempts it, instead of the prior fire-and-forget silent-loss pattern. `__testing` export (`receiveIntoStock`, `findOrCreateSubstituteStock`) for integration tests. |
 | premadeBouquetService.js | Build / edit / dissolve / sell Premade Bouquets. Flag-on path (STOCK_Y_MODEL=true, issue #285 + #330) uses reservation model — `premade_bouquet_lines` are the ledger; Batch quantity is touched only at sale via standard allocation. Legacy path preserved under `_*Legacy` functions. Persistence via `premadeBouquetRepo` (Postgres, Phase 7). |
 | notifications.js | SSE broadcast to all connected clients (heartbeat every 30s). No catch-up buffer yet — clients miss events while disconnected. |
 | wix.js | Wix webhook processor — parses payload, creates order + lines + delivery. |
